Exercise 6.8 solution
