+++ anecdotes-redux-main/src/components/AnecdoteList.js
@@ -1,0 +1,31 @@
+import React from "react";
+import {useDispatch, useSelector} from "react-redux";
+import {voteAnecdote} from "../reducers/anecdoteReducer";
+
+const AnecdoteList = () => {
+  const dispatch = useDispatch()
+  const anecdotes = useSelector(state => state)
+
+  const vote = (id) => {
+    console.log('vote', id)
+    dispatch(voteAnecdote(id))
+  }
+
+  return (
+    <>
+      {anecdotes.sort((anecdote1, anecdote2) => anecdote2.votes - anecdote1.votes).map(anecdote =>
+        <div key={anecdote.id}>
+          <div>
+            {anecdote.content}
+          </div>
+          <div>
+            has {anecdote.votes}
+            <button onClick={() => vote(anecdote.id)}>vote</button>
+          </div>
+        </div>
+      )}
+    </>
+  )
+}
+
+export default AnecdoteList
